+++ default.nix
@@ -19,19 +19,19 @@ buildGoPackage rec {
 
         nix-prefetch-git
         (import ./nix/deps2nix {inherit stdenv fetchFromGitHub buildGoPackage;})
-
         # node for tests
         nodejs-8_x
 
         # for building releases
         openssl upx
+
+        # for signing windows releases
+        (import ./nix/osslsigncode {inherit stdenv fetchurl openssl curl autoconf;})
         # for deployment to S3
         awscli
 
-        autoconf automake libtool flex
-
-        pkgconfig
-
+        # Required for building go dependencies
+        autoconf automake libtool flex pkgconfig
       ]
       # PCSC on Darwin
       ++ lib.optional stdenv.isDarwin pkgs.darwin.apple_sdk.frameworks.PCSC